requirements.txt
annotated-types==0.7.0
anyio==4.13.0
appwrite==18.0.0
certifi==2026.4.22
charset-normalizer==3.4.7
elevenlabs==2.44.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
requests==2.33.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
